analysis-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-multipart==0.0.6
pdfplumber==0.10.3
python-docx==1.1.0
textblob==0.17.1
spacy==3.7.2
